requirements.txt
flask
flask-cors
mysql-connector-python
werkzeug
gunicorn
python-dotenv
cloudinary
pandas
openpyxl
